requirements.txt
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0 